[REL] Migration sale_multi_fix to 11.0
* [FIX] update manifest & readme
* [FIX] separe in 1 file each class of the model
* [FIX] review all funcionalty of the module if it's necesary in v11
