+++ sale_multic_fix/models/__init__.py
@@ -3,3 +3,4 @@
 # directory
 ##############################################################################
 from . import sale_order
+from . import sale_order_line
