+++ sale_multic_fix/wizards/__init__.py
@@ -2,5 +2,4 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import models
-from . import wizards
+from . import sale_advance_payment_inv
